투표 구현 완료

+++ src/main/java/com/clubmanager/service/PollServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.clubmanager.domain.PersonalRecordDTO;
+import com.clubmanager.domain.PollMoMVO;
 import com.clubmanager.domain.PollPartVO;
 import com.clubmanager.domain.PollStatusDTO;
 import com.clubmanager.mapper.PollMapper;
@@ -32,10 +33,10 @@ public class PollServiceImpl implements PollService {
 	}
 	
 	@Override
-	public List<PollStatusDTO> getPSList(PollPartVO ppVO) {
+	public List<PollStatusDTO> getPSList(int matchNo, int pollType) {
 		List<PollStatusDTO> psDTOList = new ArrayList<>();
 		
-		psDTOList = pollMapper.getPSList(ppVO);
+		psDTOList = pollMapper.getPSList(matchNo, pollType);
 		return psDTOList;
 	}
 	
@@ -49,15 +50,21 @@ public class PollServiceImpl implements PollService {
 	public boolean vote(PollStatusDTO psDTO) {
 		int result = pollMapper.modifyPS(psDTO);
 		
-		PersonalRecordDTO prDTO = new PersonalRecordDTO();
-		prDTO.setMatchNo(psDTO.getMatchNo());
-		prDTO.setUserId(psDTO.getUserId());
-		if(psDTO.getStatus()==1) {//psDTO.getStatus == 1 => 참석자 명단 추가
-			log.warn("insertPR  FROM vote..... prDTO : " + prDTO);
-			recordMapper.insertPR(prDTO);
-		}else if(psDTO.getStatus()==2) {//psDTO.getStatus == 2 => 참석자 명단 제거
-			log.warn("deletePR  FROM vote..... prDTO : " + prDTO);
-			recordMapper.deletePR(prDTO);
+		if(psDTO.getPollType()==1) {//참석 투표
+			PersonalRecordDTO prDTO = new PersonalRecordDTO();
+			prDTO.setMatchNo(psDTO.getMatchNo());
+			prDTO.setUserId(psDTO.getUserId());
+			if(psDTO.getStatus()==1) {//psDTO.getStatus == 1 => 참석자 명단 추가
+				log.warn("insertPR  FROM vote..... prDTO : " + prDTO);
+				recordMapper.insertPR(prDTO);
+			}else if(psDTO.getStatus()==2) {//psDTO.getStatus == 2 => 참석자 명단 제거
+				log.warn("deletePR  FROM vote..... prDTO : " + prDTO);
+				recordMapper.deletePR(prDTO);
+			}
+		}else {//MoM 투표
+			psDTO.setUserId(psDTO.getVoteTo());
+			log.warn("picked  FROM vote..... psDTO : " + psDTO);
+			pollMapper.picked(psDTO);
 		}
 		
 		if(result == 1) return true;
@@ -67,6 +74,18 @@ public class PollServiceImpl implements PollService {
 	@Override
 	public boolean modifyPP(PollPartVO ppVO) {
 		int result = pollMapper.modifyPP(ppVO);
+		if(result == 1) return true;
+		return false;
+	}
+	
+	@Override
+	public PollMoMVO getPM(String clubCode) {
+		return pollMapper.getPM(clubCode);
+	}
+	
+	@Override
+	public boolean modifyPM(PollMoMVO pmVO) {
+		int result = pollMapper.modifyPM(pmVO);
 		if(result == 1) return true;
 		return false;
 	}
